Revert "Use `unsignedBigInteger` for `interactions.user_id` user rela…

+++ database/migrations/2018_06_14_000000_create_acquaintances_interactions_table.php
@@ -16,7 +16,7 @@ class CreateAcquaintancesInteractionsTable extends Migration
             $userModel = config('auth.providers.users.model');
             $userModel = (new $userModel);
 
-            $table->unsignedBigInteger('user_id')->index();
+            $table->unsignedInteger('user_id')->index();
             $table->morphs('subject');
             $table->string('relation')->default('follow')->comment('follow/like/subscribe/favorite/upvote/downvote');
             $table->integer('relation_value')->nullable();
